+++ src/findex/findex.ts
@@ -5,7 +5,6 @@ import init, {
 
 import { SymmetricKey } from "../kms/structs/objects"
 import { parse as parseUuid, stringify as stringifyUuid } from "uuid"
-import { encode, decode } from "../utils/leb128"
 import { bytesEquals, hexEncode } from "../utils/utils"
 import { fromByteArray } from "base64-js"
 
@@ -91,7 +90,10 @@ export class Location {
   }
 
   static fromNumber(value: number): Location {
-    return new Location(encode(value))
+    const buffer = new ArrayBuffer(8)
+    new DataView(buffer).setBigInt64(0, BigInt(value), false)
+
+    return new Location(new Uint8Array(buffer))
   }
 
   static fromUuid(value: string): Location {
@@ -103,16 +105,13 @@ export class Location {
   }
 
   toNumber(): number {
-    const { result, tail } = decode(this.bytes)
-    if (tail.length !== 0) {
+    if (this.bytes.length !== 8) {
       throw new Error(
-        `The value encoded inside this location is not a LEB128 number created with the \`Location.fromNumber()\`. Here is the hex encoded value: ${hexEncode(
-          this.bytes,
-        )}`,
+        `The location is of length ${this.bytes.length}, 8 bytes expected for a number.`,
       )
     }
 
-    return result
+    return Number(new DataView(this.bytes.buffer).getBigInt64(0, false))
   }
 
   toUuidString(): string {